+++ resources/views/components/accordion.blade.php
@@ -1,4 +1,4 @@
-@props(['type' => '', 'label' => ''])
+@props(['type' => '', 'label' => '', 'labelSlot' => ''])
 <!-- Accordion Wrapper -->
 <div x-data="{show:false}" {{$attributes->class("transition hover:bg-indigo-50 border-b border-b-black")}}
      :class="show ? 'bg-indigo-50':''"
@@ -9,7 +9,11 @@
     <div x-on:click="show= !show"
         class="cursor-pointer transition flex space-x-5 px-5 items-center h-12">
         <i class="fas " :class="show ? 'fa-caret-up':'fa-caret-down'"></i>
+        @if($labelSlot)
+            {{$labelSlot}}
+        @else
         <h3>{{__($label)}}</h3>
+        @endif
     </div>
     <!-- Content -->
     <div x-show="show" class="px-5 pt-0 overflow-hidden" x-collapse>